requirements.txt
deta==1.1.0
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
pyodbc==4.0.35
python-dotenv==0.21.0
pywin32==305
smartsheet_python_sdk==3.0.0
SQLAlchemy==1.4.45
st_clickable_images==0.0.3
streamlit==1.16.0
streamlit_autorefresh==0.0.1
streamlit_extras==0.2.4
streamlit_multipage==0.0.18
streamlit_permalink==0.3.0
streamlit_ws_localstorage==1.0.0
websockets==10.4
~treamlit==1.15.2
